Rename ambiguous source files
