+++ src/main.ts
@@ -1,7 +1,7 @@
 'use strict';
 
-import {drawTable, addPage} from './painter';
-import {calculateWidths} from './calculator';
+import {drawTable, addPage} from './tableDrawer';
+import {calculateWidths} from './widthCalculator';
 import {parseInput} from './inputParser';
 import {setDefaults, setupState, resetState} from './state';
 import './autoTableText';
