+++ src/widthCalculator.ts
